Move functions around to avoid circular import error
