+++ cg/meta/demultiplex/demux_post_processing.py
@@ -11,11 +11,11 @@ from cg.apps.demultiplex.demux_report import create_demux_report
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.constants.cgstats import STATS_HEADER
 from cg.constants.demultiplexing import DemultiplexingDirsAndFiles
-from cg.exc import FlowCellError
+from cg.exc import FlowCellError, MissingFilesError
 from cg.meta.demultiplex import files
 from cg.meta.demultiplex.housekeeper_storage_functions import (
-    store_flow_cell_data_in_housekeeper,
     get_sample_sheets_from_latest_version,
+    store_flow_cell_data_in_housekeeper,
 )
 from cg.meta.demultiplex.status_db_storage_functions import (
     store_flow_cell_data_in_status_db,
@@ -122,7 +122,7 @@ class DemuxPostProcessingAPI:
                 flow_cell=parsed_flow_cell,
                 force=force,
             )
-        except FlowCellError as e:
+        except (FlowCellError, MissingFilesError) as e:
             LOG.error(f"Flow cell {flow_cell_directory_name} will be skipped: {e}")
             return
