+++ cg/meta/demultiplex/utils.py
@@ -6,7 +6,6 @@ from typing import List, Optional
 from cg.constants.constants import FileExtensions
 from cg.constants.demultiplexing import DemultiplexingDirsAndFiles
 from cg.constants.sequencing import FLOWCELL_Q30_THRESHOLD, Sequencers
-from cg.meta.demultiplex.validation import is_valid_sample_fastq_file
 from cg.models.demultiplex.flow_cell import FlowCellDirectoryData
 from cg.utils.files import (
     get_file_in_directory,
@@ -31,6 +30,58 @@ def get_lane_from_sample_fastq(sample_fastq_path: Path) -> int:
         return int(lane_match.group(1))
 
     raise ValueError(f"Could not extract lane number from fastq file name {sample_fastq_path.name}")
+
+
+def is_sample_id_in_directory_name(directory: Path, sample_internal_id: str) -> bool:
+    """Validate that directory name is formatted as Sample_<sample_id> or Sample_<sample_id>_."""
+    sample_pattern: str = f"Sample_{sample_internal_id}"
+    return f"{sample_pattern}_" in directory.name or sample_pattern == directory.name
+
+
+def is_sample_id_in_file_name(sample_fastq: Path, sample_internal_id: str) -> bool:
+    """Validate that file name contains the sample id formatted as <sample_id>_."""
+    return f"{sample_internal_id}_" in sample_fastq.name
+
+
+def is_file_path_compressed_fastq(file_path: Path) -> bool:
+    return file_path.name.endswith(f"{FileExtensions.FASTQ}{FileExtensions.GZIP}")
+
+
+def is_lane_in_fastq_file_name(sample_fastq: Path) -> bool:
+    """Validate that fastq contains lane number formatted as _L<lane_number>"""
+    return bool(re.search(r"_L\d+", sample_fastq.name))
+
+
+def is_valid_sample_fastq_file(sample_fastq: Path, sample_internal_id: str) -> bool:
+    """
+    Validate that the sample fastq file name is formatted as expected.
+
+    Assumptions:
+    1. The sample fastq file name ends with .fastq.gz
+    2. The sample fastq file name contains the lane number formatted as _L<lane_number>
+    3. The sample internal id is present in the parent directory name or in the file name.
+    """
+    sample_id_in_directory: bool = is_sample_id_in_directory_name(
+        directory=sample_fastq.parent, sample_internal_id=sample_internal_id
+    )
+    sample_id_in_file_name: bool = is_sample_id_in_file_name(
+        sample_fastq=sample_fastq, sample_internal_id=sample_internal_id
+    )
+
+    return (
+        is_file_path_compressed_fastq(sample_fastq)
+        and is_lane_in_fastq_file_name(sample_fastq)
+        and (sample_id_in_directory or sample_id_in_file_name)
+    )
+
+
+def get_valid_sample_fastqs(fastq_paths: List[Path], sample_internal_id: str) -> List[Path]:
+    """Return a list of valid fastq files."""
+    return [
+        fastq
+        for fastq in fastq_paths
+        if is_valid_sample_fastq_file(sample_fastq=fastq, sample_internal_id=sample_internal_id)
+    ]
 
 
 def get_sample_fastqs_from_flow_cell(
@@ -73,15 +124,6 @@ def get_sample_fastqs_from_flow_cell(
             return valid_sample_fastqs
 
 
-def get_valid_sample_fastqs(fastq_paths: List[Path], sample_internal_id: str) -> List[Path]:
-    """Return a list of valid fastq files."""
-    return [
-        fastq
-        for fastq in fastq_paths
-        if is_valid_sample_fastq_file(sample_fastq=fastq, sample_internal_id=sample_internal_id)
-    ]
-
-
 def create_delivery_file_in_flow_cell_directory(flow_cell_directory: Path) -> None:
     Path(flow_cell_directory, DemultiplexingDirsAndFiles.DELIVERY).touch()
 
@@ -106,7 +148,7 @@ def parse_flow_cell_directory_data(
 
 
 def add_flow_cell_name_to_fastq_file_path(fastq_file_path: Path, flow_cell_name: str) -> Path:
-    """Add the flow cell name to the fastq file path if the flow cell name is not already in the given file path."""
+    """Add the flow cell name to the fastq file path if missing."""
     if is_pattern_in_file_path_name(file_path=fastq_file_path, pattern=flow_cell_name):
         LOG.debug(
             f"Flow cell name {flow_cell_name} already in {fastq_file_path}. Skipping renaming."
