+++ cg/meta/demultiplex/validation.py
@@ -1,55 +1,16 @@
-import re
-from pathlib import Path
 import logging
-from cg.constants.constants import FileExtensions
+from pathlib import Path
+from typing import List, Optional
+
+from cg.apps.demultiplex.sample_sheet.read_sample_sheet import (
+    get_sample_internal_ids_from_sample_sheet,
+)
 from cg.constants.demultiplexing import DemultiplexingDirsAndFiles
-from cg.exc import FlowCellError
+from cg.exc import FlowCellError, MissingFilesError
+from cg.meta.demultiplex.utils import get_sample_fastqs_from_flow_cell
 from cg.models.demultiplex.flow_cell import FlowCellDirectoryData
 
 LOG = logging.getLogger(__name__)
-
-
-def is_valid_sample_fastq_file(sample_fastq: Path, sample_internal_id: str) -> bool:
-    """
-    Validate that the sample fastq file name is formatted as expected.
-
-    Assumptions:
-    1. The sample fastq file name ends with .fastq.gz
-    2. The sample fastq file name contains the lane number formatted as _L<lane_number>
-    3. The sample internal id is present in the parent directory name or in the file name.
-    """
-    sample_id_in_directory: bool = is_sample_id_in_directory_name(
-        directory=sample_fastq.parent, sample_internal_id=sample_internal_id
-    )
-    sample_id_in_file_name: bool = is_sample_id_in_file_name(
-        sample_fastq=sample_fastq, sample_internal_id=sample_internal_id
-    )
-
-    return (
-        is_file_path_compressed_fastq(sample_fastq)
-        and is_lane_in_fastq_file_name(sample_fastq)
-        and (sample_id_in_directory or sample_id_in_file_name)
-    )
-
-
-def is_file_path_compressed_fastq(file_path: Path) -> bool:
-    return file_path.name.endswith(f"{FileExtensions.FASTQ}{FileExtensions.GZIP}")
-
-
-def is_lane_in_fastq_file_name(sample_fastq: Path) -> bool:
-    """Validate that fastq contains lane number formatted as _L<lane_number>"""
-    return bool(re.search(r"_L\d+", sample_fastq.name))
-
-
-def is_sample_id_in_directory_name(directory: Path, sample_internal_id: str) -> bool:
-    """Validate that directory name contains the sample id formatted as Sample_<sample_id> or Sample_<sample_id>_."""
-    sample_pattern: str = f"Sample_{sample_internal_id}"
-    return f"{sample_pattern}_" in directory.name or sample_pattern == directory.name
-
-
-def is_sample_id_in_file_name(sample_fastq: Path, sample_internal_id: str) -> bool:
-    """Validate that file name contains the sample id formatted as <sample_id>_."""
-    return f"{sample_internal_id}_" in sample_fastq.name
 
 
 def is_demultiplexing_complete(flow_cell_directory: Path) -> bool:
@@ -77,9 +38,26 @@ def validate_demultiplexing_complete(flow_cell_output_directory: Path) -> None:
 def validate_flow_cell_delivery_status(flow_cell_output_directory: Path, force: bool) -> None:
     if is_flow_cell_ready_for_delivery(flow_cell_output_directory) and not force:
         raise FlowCellError(
-            f"Flow cell output directory {flow_cell_output_directory} has already been processed and is ready for delivery."
+            f"Flow cell output directory {flow_cell_output_directory}"
+            "has already been processed and is ready for delivery."
         )
-    return
+
+
+def validate_samples_have_fastq_files(flow_cell: FlowCellDirectoryData) -> None:
+    """Check if all samples have already all the fastq files in the demultiplex directory.
+    Raises: MissingFilesError
+        When one of the samples does not have enough fastq files in the flow cell
+    """
+    sample_ids: List[str] = get_sample_internal_ids_from_sample_sheet(
+        sample_sheet_path=flow_cell.get_sample_sheet_path_hk(),
+        flow_cell_sample_type=flow_cell.sample_type,
+    )
+    for sample_id in sample_ids:
+        fastq_files: Optional[List[Path]] = get_sample_fastqs_from_flow_cell(
+            flow_cell_directory=flow_cell.path, sample_internal_id=sample_id
+        )
+        if not fastq_files:
+            raise MissingFilesError(f"Sample {sample_id} has no fastq files in flow cell")
 
 
 def is_flow_cell_ready_for_postprocessing(
@@ -89,4 +67,7 @@ def is_flow_cell_ready_for_postprocessing(
 ) -> None:
     validate_sample_sheet_exists(flow_cell)
     validate_demultiplexing_complete(flow_cell_output_directory)
-    validate_flow_cell_delivery_status(flow_cell_output_directory, force=force)
+    validate_flow_cell_delivery_status(
+        flow_cell_output_directory=flow_cell_output_directory, force=force
+    )
+    validate_samples_have_fastq_files(flow_cell)
